Fix some warnings.
git-svn-id: http://eathena.googlecode.com/svn/trunk@15153 9c4de8d1-4af9-ca9b-867f-b47ac08f6c80

+++ src/map/mapreg_sql.c
@@ -24,7 +24,7 @@ static bool mapreg_dirty = false;
 /// Looks up the value of an integer variable using its uid.
 int mapreg_readreg(int uid)
 {
-	return (int)idb_get(mapreg_db, uid);
+	return (int)(intptr_t)idb_get(mapreg_db, uid);
 }
 
 /// Looks up the value of a string variable using its uid.
@@ -42,7 +42,7 @@ bool mapreg_setreg(int uid, int val)
 
 	if( val != 0 )
 	{
-		if( idb_put(mapreg_db,uid,(void*)val) )
+		if( idb_put(mapreg_db,uid,(void*)(intptr_t)val) )
 			mapreg_dirty = true; // already exists, delay write
 		else if(name[1] != '@')
 		{// write new wariable to database
@@ -134,7 +134,7 @@ static void script_load_mapreg(void)
 		if( varname[length-1] == '$' )
 			idb_put(mapregstr_db, (i<<24)|s, aStrdup(value));
 		else
-			idb_put(mapreg_db, (i<<24)|s, (void *)atoi(value));
+			idb_put(mapreg_db, (i<<24)|s, (void *)(intptr_t)atoi(value));
 	}
 	
 	SqlStmt_Free(stmt);
@@ -159,7 +159,7 @@ static void script_save_mapreg(void)
 		if( name[1] == '@' )
 			continue;
 
-		if( SQL_ERROR == Sql_Query(mmysql_handle, "UPDATE `%s` SET `value`='%d' WHERE `varname`='%s' AND `index`='%d'", mapreg_table, (int)data, name, i) )
+		if( SQL_ERROR == Sql_Query(mmysql_handle, "UPDATE `%s` SET `value`='%d' WHERE `varname`='%s' AND `index`='%d'", mapreg_table, (int)(intptr_t)data, name, i) )
 			Sql_ShowDebug(mmysql_handle);
 	}
 	iter->destroy(iter);
